[CES-3190] Update the CMake/CPack to allow versioning
With this commit we update the CMakeLists and CPackConfig to allow
setting a package version via the command line when building.

This repository is rarely touched so it has no CI infrastructure and
also is legacy code. Investing in adding CI infrastructure might be
unecessary since we might also be able to just source the assets
directly from the swagger0ui offical github.

The structure of this ultimaker version of swagger-ui seems stripped
down to the bare minimum and seems slightly customized.

The README.md has been updated with some steps on how to build the
package from the command line.